AI_agent_expert_system/frontend/requirements.txt
streamlit>=1.30.0
httpx>=0.25.0
python-dotenv>=1.0.0
streamlit-mermaid>=0.1.0
plotly>=5.18.0
pandas>=2.0.0
